feat: let consumer encode pda seeds
